feat(Exec): define officer model
