+++ src/lib/models/board.ts
@@ -1,15 +1,5 @@
 import * as v from 'valibot';
-
-export const POSITIONS = [
-    'President',
-    'Vice President for Innovation',
-    'Vice President for Service',
-    'Director for External Relations',
-    'Director for Membership & Internals',
-    'Director for Branding & Creatives',
-    'Director for Engineering',
-] as const;
-
+import { Officer } from './officer';
 
 export interface Board extends v.InferOutput<typeof Board> {
     src: string;
@@ -17,8 +7,5 @@ export interface Board extends v.InferOutput<typeof Board> {
 
 export const Board = v.object({
     year: v.string(),
-    officers: v.array(v.record(
-        v.string(),
-        v.picklist(POSITIONS),
-    )),
+    officers: v.array(Officer),
 });
